+++ imagemeta/svg.go
@@ -1,4 +1,4 @@
-package imagesize
+package imagemeta
 
 import (
 	"io"
@@ -7,8 +7,8 @@ import (
 func init() {
 	// Register fake svg decoder. Since we need this only for type detecting, we can
 	// return fake image sizes
-	decodeMeta := func(io.Reader) (*Meta, error) {
-		return &Meta{Format: "svg", Width: 1, Height: 1}, nil
+	decodeMeta := func(io.Reader) (Meta, error) {
+		return &meta{format: "svg", width: 1, height: 1}, nil
 	}
 	RegisterFormat("<?xml ", decodeMeta)
 	RegisterFormat("<svg", decodeMeta)