+++ processing_options.go
@@ -12,7 +12,7 @@ import (
 	"strings"
 	"sync"
 
-	structdiff "github.com/imgproxy/imgproxy/struct-diff"
+	"github.com/imgproxy/imgproxy/structdiff"
 )
 
 type urlOption struct {
@@ -93,6 +93,11 @@ type gravityOptions struct {
 	X, Y float64
 }
 
+type extendOptions struct {
+	Enabled bool
+	Gravity gravityOptions
+}
+
 type cropOptions struct {
 	Width   int
 	Height  int
@@ -103,9 +108,7 @@ type watermarkOptions struct {
 	Enabled   bool
 	Opacity   float64
 	Replicate bool
-	Gravity   gravityType
-	OffsetX   int
-	OffsetY   int
+	Gravity   gravityOptions
 	Scale     float64
 }
 
@@ -116,7 +119,7 @@ type processingOptions struct {
 	Dpr          float64
 	Gravity      gravityOptions
 	Enlarge      bool
-	Extend       bool
+	Extend       extendOptions
 	Crop         cropOptions
 	Format       imageType
 	Quality      int
@@ -144,8 +147,9 @@ const (
 	urlTokenPlain           = "plain"
 	maxClientHintDPR        = 8
 
-	msgForbidden  = "Forbidden"
-	msgInvalidURL = "Invalid URL"
+	msgForbidden     = "Forbidden"
+	msgInvalidURL    = "Invalid URL"
+	msgInvalidSource = "Invalid Source"
 )
 
 func (gt gravityType) String() string {
@@ -197,6 +201,7 @@ func newProcessingOptions() *processingOptions {
 			Height:       0,
 			Gravity:      gravityOptions{Type: gravityCenter},
 			Enlarge:      false,
+			Extend:       extendOptions{Enabled: false, Gravity: gravityOptions{Type: gravityCenter}},
 			Quality:      conf.Quality,
 			MaxBytes:     0,
 			Format:       imageTypeUnknown,
@@ -204,7 +209,7 @@ func newProcessingOptions() *processingOptions {
 			Blur:         0,
 			Sharpen:      0,
 			Dpr:          1,
-			Watermark:    watermarkOptions{Opacity: 1, Replicate: false, Gravity: gravityCenter},
+			Watermark:    watermarkOptions{Opacity: 1, Replicate: false, Gravity: gravityOptions{Type: gravityCenter}},
 		}
 	})
 
@@ -419,17 +424,27 @@ func applyEnlargeOption(po *processingOptions, args []string) error {
 }
 
 func applyExtendOption(po *processingOptions, args []string) error {
-	if len(args) > 1 {
+	if len(args) > 4 {
 		return fmt.Errorf("Invalid extend arguments: %v", args)
 	}
 
-	po.Extend = parseBoolOption(args[0])
+	po.Extend.Enabled = parseBoolOption(args[0])
+
+	if len(args) > 1 {
+		if err := parseGravity(&po.Extend.Gravity, args[1:]); err != nil {
+			return err
+		}
+
+		if po.Extend.Gravity.Type == gravitySmart {
+			return errors.New("extend doesn't support smart gravity")
+		}
+	}
 
 	return nil
 }
 
 func applySizeOption(po *processingOptions, args []string) (err error) {
-	if len(args) > 4 {
+	if len(args) > 7 {
 		return fmt.Errorf("Invalid size arguments: %v", args)
 	}
 
@@ -451,8 +466,8 @@ func applySizeOption(po *processingOptions, args []string) (err error) {
 		}
 	}
 
-	if len(args) == 4 && len(args[3]) > 0 {
-		if err = applyExtendOption(po, args[3:4]); err != nil {
+	if len(args) >= 4 && len(args[3]) > 0 {
+		if err = applyExtendOption(po, args[3:]); err != nil {
 			return
 		}
 	}
@@ -475,7 +490,7 @@ func applyResizingTypeOption(po *processingOptions, args []string) error {
 }
 
 func applyResizeOption(po *processingOptions, args []string) error {
-	if len(args) > 5 {
+	if len(args) > 8 {
 		return fmt.Errorf("Invalid resize arguments: %v", args)
 	}
 
@@ -667,7 +682,7 @@ func applyWatermarkOption(po *processingOptions, args []string) error {
 		if args[1] == "re" {
 			po.Watermark.Replicate = true
 		} else if g, ok := gravityTypes[args[1]]; ok && g != gravityFocusPoint && g != gravitySmart {
-			po.Watermark.Gravity = g
+			po.Watermark.Gravity.Type = g
 		} else {
 			return fmt.Errorf("Invalid watermark position: %s", args[1])
 		}
@@ -675,7 +690,7 @@ func applyWatermarkOption(po *processingOptions, args []string) error {
 
 	if len(args) > 2 && len(args[2]) > 0 {
 		if x, err := strconv.Atoi(args[2]); err == nil {
-			po.Watermark.OffsetX = x
+			po.Watermark.Gravity.X = float64(x)
 		} else {
 			return fmt.Errorf("Invalid watermark X offset: %s", args[2])
 		}
@@ -683,7 +698,7 @@ func applyWatermarkOption(po *processingOptions, args []string) error {
 
 	if len(args) > 3 && len(args[3]) > 0 {
 		if y, err := strconv.Atoi(args[3]); err == nil {
-			po.Watermark.OffsetY = y
+			po.Watermark.Gravity.Y = float64(y)
 		} else {
 			return fmt.Errorf("Invalid watermark Y offset: %s", args[3])
 		}
@@ -793,6 +808,18 @@ func applyProcessingOptions(po *processingOptions, options urlOptions) error {
 	}
 
 	return nil
+}
+
+func isAllowedSource(imageURL string) bool {
+	if len(conf.AllowedSources) == 0 {
+		return true
+	}
+	for _, val := range conf.AllowedSources {
+		if strings.HasPrefix(imageURL, string(val)) {
+			return true
+		}
+	}
+	return false
 }
 
 func parseURLOptions(opts []string) (urlOptions, []string) {
@@ -986,6 +1013,10 @@ func parsePath(ctx context.Context, r *http.Request) (context.Context, error) {
 
 	if err != nil {
 		return ctx, newError(404, err.Error(), msgInvalidURL)
+	}
+
+	if !isAllowedSource(imageURL) {
+		return ctx, newError(404, fmt.Sprintf("Invalid source"), msgInvalidSource)
 	}
 
 	ctx = context.WithValue(ctx, imageURLCtxKey, imageURL)